Добился компиляции
