--- hw04_lru_cache/cache.go
@@ -1,25 +1,0 @@
-package hw04lrucache
-
-type Key string
-
-type Cache interface {
-	Set(key Key, value interface{}) bool
-	Get(key Key) (interface{}, bool)
-	Clear()
-}
-
-type lruCache struct {
-	Cache // Remove me after realization.
-
-	capacity int
-	queue    List
-	items    map[Key]*ListItem
-}
-
-func NewCache(capacity int) Cache {
-	return &lruCache{
-		capacity: capacity,
-		queue:    NewList(),
-		items:    make(map[Key]*ListItem, capacity),
-	}
-}
