+++ hw04_lru_cache/list.go
@@ -1,4 +1,4 @@
-package hw04lrucache
+package main
 
 type List interface {
 	Len() int
