+++ hw04_lru_cache/list_test.go
@@ -1,4 +1,4 @@
-package hw04lrucache
+package main
 
 import (
 	"testing"
