+++ hw04_lru_cache/main.go
@@ -1,0 +1,33 @@
+package main
+
+import (
+	"fmt"
+)
+
+func main() {
+	fmt.Println("hw04_lru_cache - function main")
+}
+
+type Key string
+
+type Cache interface {
+	Set(key Key, value interface{}) bool
+	Get(key Key) (interface{}, bool)
+	Clear()
+}
+
+type lruCache struct {
+	Cache // Remove me after realization.
+
+	capacity int
+	queue    List
+	items    map[Key]*ListItem
+}
+
+func NewCache(capacity int) Cache {
+	return &lruCache{
+		capacity: capacity,
+		queue:    NewList(),
+		items:    make(map[Key]*ListItem, capacity),
+	}
+}
